Renames record/replay files.
Mahimahi's replayserver.cc hashes the beginning of a request to match it with
the name of the corresponding on-disk file, but it seemed like a different hash
function was used for these recordings than what's in the repo now because none of
them were being matched.  I extracted the beginning of the requests from each of
these and rehashed, then renamed according to the new hash.